requirements.txt
flask>=3.0.0
garminconnect
garth
supabase
PyJWT
cryptography
